Remove behaviors completely
