+++ docs/guide/README.md
@@ -6,7 +6,6 @@ An extension for running tasks asynchronously via queues.
 
 * [Usage basics](usage.md)
 * [Migrating from `yii2-queue`](migrating-from-yii2-queue.md)
-* [Message behaviors](behaviors.md)
 * [Errors and retryable jobs](retryable.md)
 * [Workers](worker.md)
 * [Adapter list](adapter-list.md)
