--- src/Message/Behaviors/BehaviorInterface.php
@@ -1,19 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Yiisoft\Yii\Queue\Message\Behaviors;
-
-/**
- * Interface for behaviors to be used by queue adapters.
- */
-interface BehaviorInterface
-{
-    /**
-     * Returns current state data as an array of constructor parameters.
-     * Behavior will be restored with this data after serializing and deserializing.
-     *
-     * @return array
-     */
-    public function getConstructorParameters(): array;
-}
