--- tests/App/DummyBehavior.php
@@ -1,15 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Yiisoft\Yii\Queue\Tests\App;
-
-use Yiisoft\Yii\Queue\Message\Behaviors\BehaviorInterface;
-
-class DummyBehavior implements BehaviorInterface
-{
-    public function getConstructorParameters(): array
-    {
-        return [];
-    }
-}
