--- tests/App/DummyInterface.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Yiisoft\Yii\Queue\Tests\App;
-
-interface DummyInterface
-{
-}
